--- Assets/Scripts/Business/Map/DecorPart.cs
@@ -1,8 +1,0 @@
-public class DecorPart
-{
-    private double _obstructionLevel;
-
-    public DecorPart(double obstructionLevel) { _obstructionLevel = obstructionLevel; }
-
-    public double ObstructionLevel { get => _obstructionLevel; set => _obstructionLevel = value; }
-}
