+++ RPGenerator/UserSettings/Search.settings
@@ -1,0 +1,1 @@
+{}
